MDEV-38305: Expose adaptive hash index statistics in ANALYZE FORMAT=JSON
- btr_sea::n_parts : change from ulong to uint to reduce struct size
  and eliminate padding hole

+++ storage/innobase/include/btr0sea.h
@@ -428,7 +428,7 @@ public:
   /** number of hash table entries, to be multiplied by n_parts */
   uint n_cells;
   /** innodb_adaptive_hash_index_parts */
-  ulong n_parts;
+  uint n_parts;
   /** Partitions of the adaptive hash index */
   partition parts[512];
